requirements.txt
Django[argon2]==3.2.2

# Postgres
psycopg2==2.8.6

# Improved handling of environment variables
python-decouple==3.4
